requirements.txt
django-environ~=0.4.1
Django~=1.9.0
flake8~=3.2.1
gunicorn~=19.6.0
pyjade~=4.0.0
requests-oauthlib~=0.7.0
django-crispy-forms~=1.6.1
django-safespace~=0.2.0
